+++ api/release-notes.md
@@ -14,10 +14,10 @@ Welcome to the v8.8.0 release of the Opentrons robot software! This release incl
 
 ### New Features
 
-- Use new non-blocking commands for the Heater-Shaker, Temperature, and Thermocycler Modules. These commands can control multiple modules and complete pipetting actions simultaneously. 
+- Use new concurrent commands for the Heater-Shaker, Temperature, and Thermocycler Modules. These commands can control multiple modules and complete pipetting actions simultaneously. 
 -  Dynamic liquid tracking lets you aspirate, dispense, or mix at the liquid meniscus. Flex pipettes can track the liquid meniscus as its position changes during a pipetting action. 
 - Capture images of the Flex or OT-2 deck during a protocol. 
-- Choose where your Flex pipette will select a new tip when perfomring a transfer with a liquid class.
+- Choose where your Flex pipette will select a new tip when performing a transfer with a liquid class.
 
 ---